Crate tray_icon

tray-icon lets you create tray icons for desktop applications.

Platforms supported:

Windows

macOS

Linux (gtk Only)

Platform-specific notes:

On Windows and Linux, an event loop must be running on the thread, on Windows, a win32 event loop and on Linux, a gtk event loop. It doesn’t need to be the main thread but you have to create the tray icon on the same thread as the event loop.

On macOS, an event loop must be running on the main thread so you also need to create the tray icon on the main thread. You must make sure that the event loop is already running and not just created before creating a TrayIcon to prevent issues with fullscreen apps. In Winit for example the earliest you can create icons is on StartCause::Init.

Dependencies (Linux Only)

On Linux, gtk, libxdo is used to make the predfined Copy, Cut, Paste and SelectAll menu items work and libappindicator or libayatnat-appindicator are used to create the tray icon, so make sure to install them on your system.

Arch Linux / Manjaro:

pacman -S gtk3 xdotool libappindicator-gtk3 #or libayatana-appindicator

Debian / Ubuntu:

sudo apt install libgtk-3-dev libxdo-dev libappindicator3-dev #or libayatana-appindicator3-dev

Examples

Create a tray icon without a menu.

use tray_icon::{TrayIconBuilder, Icon};

let tray_icon = TrayIconBuilder::new()
    .with_tooltip("system-tray - tray icon library!")
    .with_icon(icon)
    .build()
    .unwrap();

Create a tray icon with a menu.

use tray_icon::{TrayIconBuilder, menu::Menu,Icon};

let tray_menu = Menu::new();
let tray_icon = TrayIconBuilder::new()
    .with_menu(Box::new(tray_menu))
    .with_tooltip("system-tray - tray icon library!")
    .with_icon(icon)
    .build()
    .unwrap();

Processing tray events

You can use TrayIconEvent::receiver to get a reference to the TrayIconEventReceiver which you can use to listen to events when a click happens on the tray icon

use tray_icon::TrayIconEvent;

if let Ok(event) = TrayIconEvent::receiver().try_recv() {
    println!("{:?}", event);
}

You can also listen for the menu events using MenuEvent::receiver to get events for the tray context menu.

use tray_icon::{TrayIconEvent, menu::MenuEvent};

if let Ok(event) = TrayIconEvent::receiver().try_recv() {
    println!("tray event: {:?}", event);
}

if let Ok(event) = MenuEvent::receiver().try_recv() {
    println!("menu event: {:?}", event);
}

Note for winit or tao users:

You should use TrayIconEvent::set_event_handler and forward the tray icon events to the event loop by using EventLoopProxy so that the event loop is awakened on each tray icon event. Same can be done for menu events using [MenuEvent::set_event_handler].

enum UserEvent {
  TrayIconEvent(tray_icon::TrayIconEvent),
  MenuEvent(tray_icon::menu::MenuEvent)
}

let event_loop = EventLoop::<UserEvent>::with_user_event().build().unwrap();

let proxy = event_loop.create_proxy();
tray_icon::TrayIconEvent::set_event_handler(Some(move |event| {
    proxy.send_event(UserEvent::TrayIconEvent(event));
}));

let proxy = event_loop.create_proxy();
tray_icon::menu::MenuEvent::set_event_handler(Some(move |event| {
    proxy.send_event(UserEvent::MenuEvent(event));
}));

Re-exports

pub use muda::dpi;

Modules

menu

Re-export of muda crate and used for tray context menu.

Structs

Icon

An icon used for the window titlebar, taskbar, etc.

Rect

Describes a rectangle including position (x - y axis) and size.

TrayIcon

Tray icon struct and associated methods.

TrayIconAttributes

Attributes to use when creating a tray icon.

TrayIconBuilder

TrayIcon builder struct and associated methods.

TrayIconId

An unique id that is associated with a tray icon.

Enums

BadIcon

An error produced when using Icon::from_rgba with invalid arguments.

Error

Errors returned by tray-icon.

MouseButton

Describes which mouse button triggered the event..

MouseButtonState

Describes the mouse button state.

TrayIconEvent

Describes a tray icon event.

Type Aliases

Result

Convenient type alias of Result type for tray-icon.

TrayIconEventReceiver

A reciever that could be used to listen to tray events.