tray_icon

Struct TrayIcon

pub struct TrayIcon { /* private fields */ }

Tray icon struct and associated methods.

This type is reference-counted and the icon is removed when the last instance is dropped.

Implementations

impl TrayIcon

pub fn new(attrs: TrayIconAttributes) -> Result<Self>

Builds and adds a new tray icon to the system tray.

Platform-specific:

Linux: Sometimes the icon wonโt be visible unless a menu is set. Setting an empty Menu is enough.

pub fn with_id<I: Into<TrayIconId>>( id: I, attrs: TrayIconAttributes, ) -> Result<Self>

Builds and adds a new tray icon to the system tray with the specified Id.

See TrayIcon::new for more info.

pub fn id(&self) -> &TrayIconId

Returns the id associated with this tray icon.

pub fn set_icon(&self, icon: Option<Icon>) -> Result<()>

Set new tray icon. If None is provided, it will remove the icon.

pub fn set_menu(&self, menu: Option<Box<dyn ContextMenu>>)

Set new tray menu.

Platform-specific:

Linux: once a menu is set it cannot be removed so None has no effect

pub fn set_tooltip<S: AsRef<str>>(&self, tooltip: Option<S>) -> Result<()>

Sets the tooltip for this tray icon.

Platform-specific:

Linux: Unsupported

pub fn set_title<S: AsRef<str>>(&self, title: Option<S>)

Sets the tooltip for this tray icon.

Platform-specific:

Linux: The title will not be shown unless there is an icon as well. The title is useful for numerical and other frequently updated information. In general, it shouldnโt be shown unless a user requests it as it can take up a significant amount of space on the userโs panel. This may not be shown in all visualizations.

Windows: Unsupported

pub fn set_visible(&self, visible: bool) -> Result<()>

Show or hide this tray icon

pub fn set_temp_dir_path<P: AsRef<Path>>(&self, path: Option<P>)

Sets the tray icon temp dir path. Linux only.

On Linux, we need to write the icon to the disk and usually it will be $XDG_RUNTIME_DIR/tray-icon or $TEMP/tray-icon.

pub fn set_icon_as_template(&self, is_template: bool)

Set the current icon as a template. macOS only.

pub fn set_icon_with_as_template( &self, icon: Option<Icon>, is_template: bool, ) -> Result<()>

pub fn set_show_menu_on_left_click(&self, enable: bool)

Disable or enable showing the tray menu on left click.

Platform-specific:

Linux: Unsupported.

pub fn rect(&self) -> Option<Rect>

Get tray icon rect.

Platform-specific:

Linux: Unsupported.