tray_icon

Struct TrayIconAttributes

pub struct TrayIconAttributes {
    pub tooltip: Option<String>,
    pub menu: Option<Box<dyn ContextMenu>>,
    pub icon: Option<Icon>,
    pub temp_dir_path: Option<PathBuf>,
    pub icon_is_template: bool,
    pub menu_on_left_click: bool,
    pub title: Option<String>,
}

Attributes to use when creating a tray icon.

Fields
tooltip: Option<String>

Tray icon tooltip

Platform-specific:

Linux: Unsupported.

menu: Option<Box<dyn ContextMenu>>

Tray menu

Platform-specific:

Linux: once a menu is set, it cannot be removed.

icon: Option<Icon>

Tray icon

Platform-specific:

Linux: Sometimes the icon won’t be visible unless a menu is set. Setting an empty Menu is enough.

temp_dir_path: Option<PathBuf>

Tray icon temp dir path. Linux only.

icon_is_template: bool

Use the icon as a template. macOS only.

menu_on_left_click: bool

Whether to show the tray menu on left click or not, default is true. macOS & Windows only.

title: Option<String>

Tray icon title.

Platform-specific

Linux: The title will not be shown unless there is an icon as well. The title is useful for numerical and other frequently updated information. In general, it shouldn’t be shown unless a user requests it as it can take up a significant amount of space on the user’s panel. This may not be shown in all visualizations.

Windows: Unsupported.